+++ com.unity.ml-agents/Runtime/Communicator/RpcCommunicator.cs
@@ -504,6 +504,20 @@ namespace MLAgents
                     "A side channel with type index {0} is already registered. You cannot register multiple " +
                     "side channels of the same id.", channelId));
             }
+
+            var numMessages = m_CachedMessages.Count;
+            for (int i = 0; i < numMessages; i++)
+            {
+                var cachedMessage = m_CachedMessages.Dequeue();
+                if (channelId == cachedMessage.ChannelId)
+                {
+                    sideChannel.OnMessageReceived(cachedMessage.Message);
+                }
+                else
+                {
+                    m_CachedMessages.Enqueue(cachedMessage);
+                }
+            }
             m_SideChannels.Add(channelId, sideChannel);
         }
 
@@ -547,6 +561,14 @@ namespace MLAgents
             }
         }
 
+        private struct CachedSideChannelMessage
+        {
+            public Guid ChannelId;
+            public byte[] Message;
+        }
+
+        private static Queue<CachedSideChannelMessage> m_CachedMessages = new Queue<CachedSideChannelMessage>();
+
         /// <summary>
         /// Separates the data received from Python into individual messages for each registered side channel.
         /// </summary>
@@ -554,6 +576,22 @@ namespace MLAgents
         /// <param name="dataReceived">The byte array of data received from Python.</param>
         public static void ProcessSideChannelData(Dictionary<Guid, SideChannel> sideChannels, byte[] dataReceived)
         {
+
+            while(m_CachedMessages.Count!=0)
+            {
+                var cachedMessage = m_CachedMessages.Dequeue();
+                if (sideChannels.ContainsKey(cachedMessage.ChannelId))
+                {
+                    sideChannels[cachedMessage.ChannelId].OnMessageReceived(cachedMessage.Message);
+                }
+                else
+                {
+                    Debug.Log(string.Format(
+                        "Unknown side channel data received. Channel Id is "
+                        + ": {0}", cachedMessage.ChannelId));
+                }
+            }
+
             if (dataReceived.Length == 0)
             {
                 return;
@@ -585,9 +623,13 @@ namespace MLAgents
                         }
                         else
                         {
-                            Debug.Log(string.Format(
-                                "Unknown side channel data received. Channel Id is "
-                                + ": {0}", channelId));
+                            // Don't recognize this ID, but cache it in case the SideChannel that can handle
+                            // it is registered before the next call to ProcessSideChannelData.
+                            m_CachedMessages.Enqueue(new CachedSideChannelMessage
+                            {
+                                ChannelId = channelId,
+                                Message = message
+                            });
                         }
                     }
                 }